+++ client/consul.go
@@ -2,14 +2,13 @@
 package client
 
 import (
-	kitconsul "github.com/go-kit/kit/sd/consul"
 	"github.com/hashicorp/consul/api"
 	"github.com/iconnor-code/cogo/cerrs"
 	"github.com/iconnor-code/cogo/core"
 )
 
 type Consul struct {
-	defaultClient kitconsul.Client
+	defaultClient *api.Client
 }
 
 func NewConsul(config core.IConfig) (*Consul, error) {
@@ -24,12 +23,11 @@ func NewConsul(config core.IConfig) (*Consul, error) {
 	if err != nil {
 		return nil, cerrs.Wrap(err)
 	}
-	defaultClient := kitconsul.NewClient(defaultConsul)
 	return &Consul{
-		defaultClient: defaultClient,
+		defaultClient: defaultConsul,
 	}, nil
 }
 
-func (c *Consul) DefaultClient() kitconsul.Client {
+func (c *Consul) DefaultClient() *api.Client {
 	return c.defaultClient
 }