+++ core/idiscovery.go
@@ -3,12 +3,13 @@ package core
 import (
 	"context"
 
-	"github.com/go-kit/kit/endpoint"
 	"google.golang.org/grpc"
 )
 
+type Endpoint func(ctx context.Context, request any) (response any, err error)
+
 type IDiscovery interface {
-	Discover(ctx context.Context, serverName, serviceName, methodName string, tags []string, resp any, opts ...grpc.DialOption) (endpoint.Endpoint, error)
+	Discover(ctx context.Context, serverName, serviceName, methodName string, tags []string, resp any, opts ...grpc.DialOption) (Endpoint, error)
 }
 
 type DiscoveryOption func(d IDiscovery) error